requirements.txt
anyio==3.3.4
asgiref==3.4.1
click==8.0.3
dnspython==2.1.0
fastapi==0.70.0
h11==0.12.0
idna==3.3
pydantic==1.8.2
pymongo==3.12.1
python-dotenv==0.19.1
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
